+++ packages/plugin-image/web/src/InPluginInput.jsx
@@ -24,12 +24,10 @@ class InPluginInput extends Component {
   className = classnames(styles.inPluginInput, this.props.className);
 
   render() {
-    const { onChange, value } = this.props;
     return (
       <input
         className={this.className}
-        value={value}
-        readOnly={!onChange}
+        value={this.props.value}
         onChange={this.onChange}
         onFocus={this.handleFocus}
         onBlur={this.handleBlur}
@@ -45,7 +43,12 @@ InPluginInput.propTypes = {
   value: PropTypes.string,
   onChange: PropTypes.func,
   setFocusToBlock: PropTypes.func,
-  setInPluginEditingMode: PropTypes.func.isRequired,
+  setInPluginEditingMode: PropTypes.func,
+};
+
+InPluginInput.defaultProps = {
+  setInPluginEditingMode: () => false,
+  setFocusToBlock: () => false,
 };
 
 export default InPluginInput;